fix duty variable

+++ tps/tp3/fw/platformio/inverter/src/pwm_gen.h
@@ -9,7 +9,7 @@ void stop_gen();
 void set_ma(float ma);
 float get_ma();
 float get_freq();
-void set_freq(int freq);
+void set_freq(float freq);
 
 
 #endif
